requirements.txt
Pillow==6.2.0
torch==1.3.1
torchvision==0.4.2
numpy==1.17.4
tqdm==4.28.1
